set up boost keybind put in movement
