+++ Assets/Scripts/PlayerScripts/movement.cs
@@ -12,8 +12,8 @@ public class Movement : MonoBehaviour
    //For future use ;)
    private PlayerInput playerInput;
 
-   //Controls movement distance
-   public float moveDistance = 2f;
+    //Controls movement distance
+    public float moveDistance = 2f;
 
    //Controls speed during movement. Used for smoothing
    //"MoveTowards" function called during update
@@ -27,12 +27,24 @@ public class Movement : MonoBehaviour
    //Keeps track of player's current vertical position on grid
    private int vertical = 0;
 
-   
-   /// <summary>
-   /// Moves player up one unit on grid if player is able
-   /// </summary>
-   /// <param name="context">context of button press associated with action</param>
-   public void Up(InputAction.CallbackContext context)
+    private InputAction boost;
+    
+    void Start()
+    {
+        // set up input action for boost
+        boost = new InputAction(
+            type: InputActionType.Button,
+            binding: "Keyboard/shift");
+
+        boost.Enable();
+    }
+
+
+    /// <summary>
+    /// Moves player up one unit on grid if player is able
+    /// </summary>
+    /// <param name="context">context of button press associated with action</param>
+    public void Up(InputAction.CallbackContext context)
    {
 
       if (context.performed)  
@@ -129,6 +141,19 @@ public class Movement : MonoBehaviour
 
       transform.position = Vector3.MoveTowards(transform.position, transVec, step);
 
-   }
+    }
+
+    private void Update()
+    {
+        // handle boosting in update so doesnt break if clicked more than once a frame
+        if (boost.WasPressedThisFrame())
+        {
+            Debug.Log("boosting");
+        }
+        else if (boost.WasReleasedThisFrame())
+        {
+            Debug.Log("stopped boosting");
+        }
+    }
 
 }
