--- Assets/Scripts/StarBoost.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 57b4cac00b3bf0a469d7a745d60bdb50
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
